+++ hal/src/main/native/sim/DigitalInternal.h
@@ -45,11 +45,11 @@ struct DigitalPort {
   uint8_t channel;
   bool configSet = false;
   bool eliminateDeadband = false;
-  double maxPwm = 0;
-  double deadbandMaxPwm = 0;
-  double centerPwm = 0;
-  double deadbandMinPwm = 0;
-  double minPwm = 0;
+  int32_t maxPwm = 0;
+  int32_t deadbandMaxPwm = 0;
+  int32_t centerPwm = 0;
+  int32_t deadbandMinPwm = 0;
+  int32_t minPwm = 0;
   std::string previousAllocation;
 };
